refactor(ao): extract agents inspect adapter
