+++ cli/cmd/ao/agents_test.go
@@ -5,7 +5,6 @@ import (
 	"bytes"
 	"encoding/json"
 	"os"
-	"os/exec"
 	"path/filepath"
 	"reflect"
 	"strings"
@@ -51,267 +50,6 @@ func TestAgentsLintCmd_Registered(t *testing.T) {
 	}
 	if cmd.Flags().Lookup("json") == nil {
 		t.Error("expected --json flag")
-	}
-}
-
-func TestParseAgentsAllowlist(t *testing.T) {
-	tests := []struct {
-		name    string
-		content string
-		want    []string
-	}{
-		{
-			name: "empty content",
-			want: []string{},
-		},
-		{
-			name:    "no markers",
-			content: "ao\nlearnings\n",
-			want:    []string{},
-		},
-		{
-			name: "basic allowlist",
-			content: `# heading
-<!-- BEGIN agents-write-surfaces-allowlist -->
-ao
-learnings
-patterns
-<!-- END agents-write-surfaces-allowlist -->
-trailing
-`,
-			want: []string{"ao", "learnings", "patterns"},
-		},
-		{
-			name: "inline comments and blanks",
-			content: `<!-- BEGIN agents-write-surfaces-allowlist -->
-
-# core
-ao   # core runtime
-
-# promoted
-learnings   # promoted artifacts
-<!-- END agents-write-surfaces-allowlist -->
-`,
-			want: []string{"ao", "learnings"},
-		},
-		{
-			name: "duplicates dedup and sorted",
-			content: `<!-- BEGIN agents-write-surfaces-allowlist -->
-patterns
-ao
-learnings
-ao
-<!-- END agents-write-surfaces-allowlist -->
-`,
-			want: []string{"ao", "learnings", "patterns"},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got := parseAgentsAllowlist(tt.content)
-			if !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("parseAgentsAllowlist() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-// TestParseAgentsAllowlist_ShellParity locks parity between the Go parser
-// (parseAgentsAllowlist) and the shell parser pipeline embedded in
-// scripts/check-agents-write-surfaces.sh. Per the L58 finding (epic ag-0af),
-// these two parsers MUST produce identical output for the same contract
-// document — otherwise the CI gate (shell) and the `ao agents inspect`
-// command (Go) can disagree about the allowlist.
-//
-// The test feeds a fixture covering every meaningful boundary the parsers
-// must agree on (BEGIN/END markers, blank lines, leading-whitespace lines,
-// inline `# comment` suffixes, full-line `# comment` lines, duplicate
-// entries, sort order) into both parsers and asserts the sorted-deduped
-// output matches. The shell pipeline is the canonical 9-line awk|sed|awk|
-// sort -u from check-agents-write-surfaces.sh; copying it here is the
-// fixture-lock — drift in either parser fails the test.
-func TestParseAgentsAllowlist_ShellParity(t *testing.T) {
-	if _, err := exec.LookPath("bash"); err != nil {
-		t.Skip("bash not available")
-	}
-	if _, err := exec.LookPath("awk"); err != nil {
-		t.Skip("awk not available")
-	}
-
-	fixture := `# AgentOps Write Surface Contract
-
-Some doc prose.
-
-<!-- BEGIN agents-write-surfaces-allowlist -->
-ao
-learnings   # promoted artifacts
-patterns
-ao
-
-# core runtime
-findings
-overnight   # dream output
-sessions
-<!-- END agents-write-surfaces-allowlist -->
-
-trailing prose.
-`
-
-	goOut := parseAgentsAllowlist(fixture)
-
-	pipeline := `awk '
-  /^[[:space:]]*<!-- BEGIN agents-write-surfaces-allowlist -->[[:space:]]*$/ { inside=1; next }
-  /^[[:space:]]*<!-- END agents-write-surfaces-allowlist -->[[:space:]]*$/   { inside=0; next }
-  inside { print }
-' \
-  | sed -E 's/[[:space:]]+#.*$//' \
-  | sed -E 's/^[[:space:]]+//; s/[[:space:]]+$//' \
-  | awk 'NF && $1 !~ /^#/' \
-  | sort -u`
-
-	cmd := exec.Command("bash", "-c", pipeline)
-	cmd.Stdin = strings.NewReader(fixture)
-	var stdout, stderr bytes.Buffer
-	cmd.Stdout = &stdout
-	cmd.Stderr = &stderr
-	if err := cmd.Run(); err != nil {
-		t.Fatalf("shell pipeline failed: %v\nstderr: %s", err, stderr.String())
-	}
-
-	rawShell := strings.Split(strings.TrimRight(stdout.String(), "\n"), "\n")
-	shellOut := []string{}
-	for _, l := range rawShell {
-		if l != "" {
-			shellOut = append(shellOut, l)
-		}
-	}
-
-	if !reflect.DeepEqual(goOut, shellOut) {
-		t.Errorf("shell-parser/Go-parser allowlist drift:\n  go:    %v\n  shell: %v", goOut, shellOut)
-	}
-
-	want := []string{"ao", "findings", "learnings", "overnight", "patterns", "sessions"}
-	if !reflect.DeepEqual(goOut, want) {
-		t.Errorf("Go parser fixture-lock failure: got %v, want %v", goOut, want)
-	}
-}
-
-func TestDiscoverActiveSkills(t *testing.T) {
-	tmp := t.TempDir()
-	for _, name := range []string{"alpha", "beta"} {
-		dir := filepath.Join(tmp, name)
-		if err := os.MkdirAll(dir, 0o755); err != nil {
-			t.Fatal(err)
-		}
-		if err := os.WriteFile(filepath.Join(dir, "SKILL.md"), []byte("ok"), 0o644); err != nil {
-			t.Fatal(err)
-		}
-	}
-	// dir without SKILL.md should be ignored
-	if err := os.MkdirAll(filepath.Join(tmp, "no-skill-md"), 0o755); err != nil {
-		t.Fatal(err)
-	}
-	// regular file should be ignored
-	if err := os.WriteFile(filepath.Join(tmp, "stray.md"), []byte("nope"), 0o644); err != nil {
-		t.Fatal(err)
-	}
-
-	got := discoverActiveSkills(tmp)
-	want := []string{"alpha", "beta"}
-	if !reflect.DeepEqual(got, want) {
-		t.Errorf("discoverActiveSkills() = %v, want %v", got, want)
-	}
-}
-
-func TestDiscoverActiveSkills_MissingDir(t *testing.T) {
-	got := discoverActiveSkills(filepath.Join(t.TempDir(), "nope"))
-	if len(got) != 0 {
-		t.Errorf("expected empty result, got %v", got)
-	}
-}
-
-func TestRunAgentsInspect_TextOutput(t *testing.T) {
-	tmp := t.TempDir()
-	contract := filepath.Join(tmp, "contract.md")
-	contractContent := `# Title
-<!-- BEGIN agents-write-surfaces-allowlist -->
-ao
-learnings
-<!-- END agents-write-surfaces-allowlist -->
-`
-	if err := os.WriteFile(contract, []byte(contractContent), 0o644); err != nil {
-		t.Fatal(err)
-	}
-
-	origJSON := agentsInspectJSON
-	origContract := agentsInspectContract
-	t.Cleanup(func() {
-		agentsInspectJSON = origJSON
-		agentsInspectContract = origContract
-	})
-	agentsInspectJSON = false
-	agentsInspectContract = contract
-
-	var buf bytes.Buffer
-	agentsInspectCmd.SetOut(&buf)
-	t.Cleanup(func() { agentsInspectCmd.SetOut(nil) })
-
-	if err := runAgentsInspect(agentsInspectCmd, nil); err != nil {
-		t.Fatalf("runAgentsInspect: %v", err)
-	}
-	out := buf.String()
-	for _, want := range []string{
-		"Contract: " + contract,
-		"Catalogued surfaces: 2",
-		".agents/ao/",
-		".agents/learnings/",
-	} {
-		if !strings.Contains(out, want) {
-			t.Errorf("output missing %q\nGot:\n%s", want, out)
-		}
-	}
-}
-
-func TestRunAgentsInspect_JSONOutput(t *testing.T) {
-	tmp := t.TempDir()
-	contract := filepath.Join(tmp, "contract.md")
-	contractContent := `<!-- BEGIN agents-write-surfaces-allowlist -->
-ao
-patterns
-<!-- END agents-write-surfaces-allowlist -->
-`
-	if err := os.WriteFile(contract, []byte(contractContent), 0o644); err != nil {
-		t.Fatal(err)
-	}
-
-	origJSON := agentsInspectJSON
-	origContract := agentsInspectContract
-	t.Cleanup(func() {
-		agentsInspectJSON = origJSON
-		agentsInspectContract = origContract
-	})
-	agentsInspectJSON = true
-	agentsInspectContract = contract
-
-	var buf bytes.Buffer
-	agentsInspectCmd.SetOut(&buf)
-	t.Cleanup(func() { agentsInspectCmd.SetOut(nil) })
-
-	if err := runAgentsInspect(agentsInspectCmd, nil); err != nil {
-		t.Fatalf("runAgentsInspect: %v", err)
-	}
-
-	var got AgentsInventory
-	if err := json.Unmarshal(buf.Bytes(), &got); err != nil {
-		t.Fatalf("output not valid JSON: %v\nGot: %s", err, buf.String())
-	}
-	if got.Contract != contract {
-		t.Errorf("Contract = %q, want %q", got.Contract, contract)
-	}
-	wantList := []string{"ao", "patterns"}
-	if !reflect.DeepEqual(got.Allowlist, wantList) {
-		t.Errorf("Allowlist = %v, want %v", got.Allowlist, wantList)
 	}
 }
 
@@ -413,29 +151,6 @@ func TestRunAgentsLint_DefaultScriptResolvesFromSubdir(t *testing.T) {
 	}
 	if got["status"] != "ok" {
 		t.Errorf("status = %q, want ok", got["status"])
-	}
-}
-
-func TestRunAgentsInspect_MissingContract(t *testing.T) {
-	origJSON := agentsInspectJSON
-	origContract := agentsInspectContract
-	t.Cleanup(func() {
-		agentsInspectJSON = origJSON
-		agentsInspectContract = origContract
-	})
-	agentsInspectJSON = false
-	agentsInspectContract = filepath.Join(t.TempDir(), "missing.md")
-
-	var buf bytes.Buffer
-	agentsInspectCmd.SetOut(&buf)
-	t.Cleanup(func() { agentsInspectCmd.SetOut(nil) })
-
-	err := runAgentsInspect(agentsInspectCmd, nil)
-	if err == nil {
-		t.Fatal("expected error for missing contract")
-	}
-	if !strings.Contains(err.Error(), "reading contract") {
-		t.Errorf("error = %v, want one mentioning 'reading contract'", err)
 	}
 }
